+++ src/heat.js
@@ -316,7 +316,7 @@
      */
 
     function renderControlTitleBar( bindingOptions ) {
-        if ( bindingOptions.showTitle || bindingOptions.showYearSelector || bindingOptions.showRefreshButton || bindingOptions.showExportButton ) {
+        if ( bindingOptions.showTitle || bindingOptions.showYearSelector || bindingOptions.showRefreshButton || bindingOptions.showExportButton || bindingOptions.showImportButton ) {
             var titleBar = createElement( bindingOptions.currentView.element, "div", "title-bar" ),
                 title = createElement( titleBar, "div", "title" );
 
@@ -348,6 +348,14 @@
 
                     renderTitleDropDownClickEvent( bindingOptions, statisticsChart, _elements_View_Statistics, _elements_View_Name_Statistics );
                 }
+            }
+
+            if ( bindingOptions.showImportButton ) {
+                var importData = createElementWithHTML( titleBar, "button", "import", _configuration.importButtonText );
+        
+                importData.onclick = function() {
+                    importFromFilesSelected( bindingOptions );
+                };
             }
 
             if ( bindingOptions.showExportButton ) {
@@ -1278,6 +1286,19 @@
         }
     }
 
+    function importFromFilesSelected( bindingOptions ) {
+        var input = createElementWithNoContainer( "input" );
+        input.type = "file";
+        input.accept = ".json";
+        input.multiple = "multiple";
+
+        input.onchange = function() {
+            importFromFiles( input.files, bindingOptions );
+        };
+
+        input.click();
+    }
+
     function importFromFiles( files, bindingOptions ) {
         var filesLength = files.length,
             filesCompleted = [],
@@ -1535,6 +1556,7 @@
         options.yearsToHide = getDefaultArray( options.yearsToHide, [] );
         options.showLessAndMoreLabels = getDefaultBoolean( options.showLessAndMoreLabels, true );
         options.showNumbersInGuide = getDefaultBoolean( options.showNumbersInGuide, false );
+        options.showImportButton = getDefaultBoolean( options.showImportButton, false );
 
         options = buildAttributeOptionMapView( options );
         options = buildAttributeOptionChartView( options );
@@ -1773,6 +1795,20 @@
      * Element Handling
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
+
+    function createElementWithNoContainer( type ) {
+        var result = null,
+            nodeType = type.toLowerCase(),
+            isText = nodeType === "text";
+
+        if ( !_elements_Type.hasOwnProperty( nodeType ) ) {
+            _elements_Type[ nodeType ] = isText ? _parameter_Document.createTextNode( _string.empty ) : _parameter_Document.createElement( nodeType );
+        }
+
+        result = _elements_Type[ nodeType ].cloneNode( false );
+
+        return result;
+    }
 
     function createElement( container, type, className, beforeNode ) {
         var result = null,
@@ -2778,6 +2814,7 @@
         _configuration.statisticsText = getDefaultString( _configuration.statisticsText, "Statistics" );
         _configuration.noStatisticsDataMessage = getDefaultString( _configuration.noStatisticsDataMessage, "There are currently no statistics to view." );
         _configuration.unknownTrendText = getDefaultString( _configuration.unknownTrendText, "Unknown" );
+        _configuration.importButtonText = getDefaultString( _configuration.importButtonText, "Import" );
     }
 
     function buildDefaultConfigurationArrays() {